+++ common/src/main/java/dev/denismasterherobrine/afterdark/mixin/MixinFireBlock.java
@@ -12,12 +12,10 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(FireBlock.class)
 public class MixinFireBlock {
-
     @Inject(method = "areBlocksAroundFlammable", at = @At("HEAD"), cancellable = true)
     private void preventGrassBurning(BlockView world, BlockPos pos, CallbackInfoReturnable<Boolean> cir) {
         if (world instanceof World) {
             FireCheck.GrassAroundCheck((World) world, pos, cir);
         }
     }
-
 }