+++ common/src/main/java/dev/denismasterherobrine/afterdark/registry/AfterdarkFeaturesRegistry.java
@@ -2,14 +2,15 @@ package dev.denismasterherobrine.afterdark.registry;
 
 import dev.denismasterherobrine.afterdark.features.*;
 import dev.denismasterherobrine.afterdark.features.configuration.*;
+import net.minecraft.block.Blocks;
 import net.minecraft.world.gen.ProbabilityConfig;
 import net.minecraft.world.gen.feature.DefaultFeatureConfig;
 import net.minecraft.world.gen.feature.DiskFeatureConfig;
 import net.minecraft.world.gen.feature.Feature;
-import net.minecraft.world.gen.feature.LargeDripstoneFeatureConfig;
 
 public class AfterdarkFeaturesRegistry {
-    public static final Feature<LargeDripstoneFeatureConfig> LARGE_BASALT_PILLAR = new LargeBasaltPillarFeature(LargeDripstoneFeatureConfig.CODEC);
+    public static final Feature<PillarFeatureConfiguration> LARGE_BASALT_PILLAR =
+            new ConfigurableDripstoneStylePillarFeature(PillarFeatureConfiguration.CODEC, PillarPlacementRules.basaltMix());
     public static final Feature<VerticalBlobConfiguration> CRYSTAL_SPIKE = new CrystalSpikeFeature(VerticalBlobConfiguration.CODEC);
     public static final Feature<SpiralConfiguration> SPIRAL = new SpiralFeature(SpiralConfiguration.CODEC);
     public static final Feature<VerticalBlobConfiguration> ADDITIVE_BLOB = new AdditiveBlobFeature(VerticalBlobConfiguration.CODEC);
@@ -17,30 +18,46 @@ public class AfterdarkFeaturesRegistry {
     public static final Feature<CatchingFallConfiguration> CATCHING_FALL = new CatchingFallFeature(CatchingFallConfiguration.CODEC);
     public static final Feature<DoubleBlockConfiguration> POND = new PondFeature(DoubleBlockConfiguration.CODEC);
     public static final Feature<VerticalBlobConfiguration> ADDITIVE_GROUND_BLOB = new AdditiveGroundBlobFeature(VerticalBlobConfiguration.CODEC);
-    public static final Feature<LargeDripstoneFeatureConfig> LARGE_CALCITE_PILLAR = new LargeCalcitePillarFeature(LargeDripstoneFeatureConfig.CODEC);
-    public static final Feature<LargeDripstoneFeatureConfig> LARGE_PRISMARINE_PILLAR = new LargePrismarinePillarFeature(LargeDripstoneFeatureConfig.CODEC);
-    public static final Feature<LargeDripstoneFeatureConfig> LARGE_DARK_PRISMARINE_PILLAR = new LargeDarkPrismarinePillarFeature(LargeDripstoneFeatureConfig.CODEC);
-    public static final Feature<LargeDripstoneFeatureConfig> LARGE_SLIME_PILLAR = new LargeSlimePillarFeature(LargeDripstoneFeatureConfig.CODEC);
-    public static final Feature<WaterloggableRandomPatchConfiguration> WATERLOGGABLE_RANDOM_PATCH_FEATURE = new WaterloggableRandomPatchFeature(WaterloggableRandomPatchConfiguration.CODEC);
+    public static final Feature<PillarFeatureConfiguration> LARGE_CALCITE_PILLAR =
+            new ConfigurableDripstoneStylePillarFeature(PillarFeatureConfiguration.CODEC, PillarPlacementRules.simpleIceFamilyStop(Blocks.CALCITE));
+    public static final Feature<PillarFeatureConfiguration> LARGE_PRISMARINE_PILLAR =
+            new ConfigurableDripstoneStylePillarFeature(PillarFeatureConfiguration.CODEC, PillarPlacementRules.simpleIceFamilyStop(Blocks.PRISMARINE));
+    public static final Feature<PillarFeatureConfiguration> LARGE_DARK_PRISMARINE_PILLAR =
+            new ConfigurableDripstoneStylePillarFeature(PillarFeatureConfiguration.CODEC, PillarPlacementRules.simpleIceFamilyStop(Blocks.DARK_PRISMARINE));
+    public static final Feature<PillarFeatureConfiguration> LARGE_SLIME_PILLAR =
+            new ConfigurableDripstoneStylePillarFeature(PillarFeatureConfiguration.CODEC, PillarPlacementRules.simpleIceFamilyStop(Blocks.SLIME_BLOCK));
+    public static final Feature<WaterloggableRandomPatchConfiguration> WATERLOGGABLE_RANDOM_PATCH_FEATURE =
+            new WaterloggableRandomPatchFeature(WaterloggableRandomPatchConfiguration.CODEC);
     public static final Feature<DefaultFeatureConfig> CAVE_KELP_FEATURE = new CaveKelpFeature(DefaultFeatureConfig.CODEC);
     public static final Feature<DefaultFeatureConfig> CAVE_PICKLE_FEATURE = new CavePickleFeature(DefaultFeatureConfig.CODEC);
-    public static final Feature<LargeDripstoneFeatureConfig> LARGE_DEEPSLATE_PILLAR = new LargeDeepslatePillarFeature(LargeDripstoneFeatureConfig.CODEC);
+    public static final Feature<PillarFeatureConfiguration> LARGE_DEEPSLATE_PILLAR =
+            new ConfigurableDripstoneStylePillarFeature(PillarFeatureConfiguration.CODEC, PillarPlacementRules.simpleIceFamilyStop(Blocks.DEEPSLATE));
     public static final Feature<ProbabilityConfig> CAVE_SEAGRASS_FEATURE = new CaveSeagrassFeature(ProbabilityConfig.CODEC);
-    public static final Feature<LargeDripstoneFeatureConfig> LARGE_OBSIDIAN_PILLAR = new LargeObsidianPillarFeature(LargeDripstoneFeatureConfig.CODEC);
+    public static final Feature<PillarFeatureConfiguration> LARGE_OBSIDIAN_PILLAR =
+            new ConfigurableDripstoneStylePillarFeature(PillarFeatureConfiguration.CODEC, PillarPlacementRules.simpleIceFamilyStop(Blocks.OBSIDIAN));
     public static final Feature<DiskFeatureConfig> SUPPORTED_DISK_FEATURE = new SupportedDiskFeature(DiskFeatureConfig.CODEC);
-    public static final Feature<LargeDripstoneFeatureConfig> LARGE_PACKED_ICE_PILLAR = new LargePackedIcePillarFeature(LargeDripstoneFeatureConfig.CODEC);
-    public static final Feature<LargeDripstoneFeatureConfig> LARGE_ICE_PILLAR = new LargeIcePillarFeature(LargeDripstoneFeatureConfig.CODEC);
-    public static final Feature<LargeDripstoneFeatureConfig> LARGE_BLUE_ICE_PILLAR = new LargeBlueIcePillarFeature(LargeDripstoneFeatureConfig.CODEC);
+    public static final Feature<PillarFeatureConfiguration> LARGE_PACKED_ICE_PILLAR =
+            new ConfigurableDripstoneStylePillarFeature(PillarFeatureConfiguration.CODEC, PillarPlacementRules.packedIceMix());
+    public static final Feature<PillarFeatureConfiguration> LARGE_ICE_PILLAR =
+            new ConfigurableDripstoneStylePillarFeature(PillarFeatureConfiguration.CODEC, PillarPlacementRules.simpleIceFamilyStop(Blocks.ICE));
+    public static final Feature<PillarFeatureConfiguration> LARGE_BLUE_ICE_PILLAR =
+            new ConfigurableDripstoneStylePillarFeature(PillarFeatureConfiguration.CODEC, PillarPlacementRules.simpleIceFamilyStop(Blocks.BLUE_ICE));
 
-    // Raw Ore Pillars
-    public static final Feature<LargeDripstoneFeatureConfig> LARGE_RAW_IRON_PILLAR = new LargeRawIronPillarFeature(LargeDripstoneFeatureConfig.CODEC);
-    public static final Feature<LargeDripstoneFeatureConfig> LARGE_RAW_GOLD_PILLAR = new LargeRawGoldPillarFeature(LargeDripstoneFeatureConfig.CODEC);
-    public static final Feature<LargeDripstoneFeatureConfig> LARGE_RAW_COPPER_PILLAR = new LargeRawCopperPillarFeature(LargeDripstoneFeatureConfig.CODEC);
+    public static final Feature<PillarFeatureConfiguration> LARGE_RAW_IRON_PILLAR =
+            new ConfigurableDripstoneStylePillarFeature(PillarFeatureConfiguration.CODEC, PillarPlacementRules.simpleStoneOnlyStop(Blocks.RAW_IRON_BLOCK));
+    public static final Feature<PillarFeatureConfiguration> LARGE_RAW_GOLD_PILLAR =
+            new ConfigurableDripstoneStylePillarFeature(PillarFeatureConfiguration.CODEC, PillarPlacementRules.simpleStoneOnlyStop(Blocks.RAW_GOLD_BLOCK));
+    public static final Feature<PillarFeatureConfiguration> LARGE_RAW_COPPER_PILLAR =
+            new ConfigurableDripstoneStylePillarFeature(PillarFeatureConfiguration.CODEC, PillarPlacementRules.simpleStoneOnlyStop(Blocks.RAW_COPPER_BLOCK));
 
-    // Ore Pillars
-    public static final Feature<LargeDripstoneFeatureConfig> LARGE_DIAMOND_ORE_PILLAR = new LargeDiamondOrePillarFeature(LargeDripstoneFeatureConfig.CODEC);
-    public static final Feature<LargeDripstoneFeatureConfig> LARGE_EMERALD_ORE_PILLAR = new LargeEmeraldOrePillarFeature(LargeDripstoneFeatureConfig.CODEC);
-    public static final Feature<LargeDripstoneFeatureConfig> LARGE_LAPIS_ORE_PILLAR = new LargeLapisOrePillarFeature(LargeDripstoneFeatureConfig.CODEC);
-    public static final Feature<LargeDripstoneFeatureConfig> LARGE_REDSTONE_ORE_PILLAR = new LargeRedstoneOrePillarFeature(LargeDripstoneFeatureConfig.CODEC);
-
+    public static final Feature<PillarFeatureConfiguration> LARGE_DIAMOND_ORE_PILLAR =
+            new ConfigurableDripstoneStylePillarFeature(PillarFeatureConfiguration.CODEC, PillarPlacementRules.simpleStoneOnlyStop(Blocks.DIAMOND_ORE));
+    public static final Feature<PillarFeatureConfiguration> LARGE_EMERALD_ORE_PILLAR =
+            new ConfigurableDripstoneStylePillarFeature(PillarFeatureConfiguration.CODEC, PillarPlacementRules.simpleStoneOnlyStop(Blocks.EMERALD_ORE));
+    public static final Feature<PillarFeatureConfiguration> LARGE_LAPIS_ORE_PILLAR =
+            new ConfigurableDripstoneStylePillarFeature(PillarFeatureConfiguration.CODEC, PillarPlacementRules.simpleStoneOnlyStop(Blocks.LAPIS_ORE));
+    public static final Feature<PillarFeatureConfiguration> LARGE_REDSTONE_ORE_PILLAR =
+            new ConfigurableDripstoneStylePillarFeature(PillarFeatureConfiguration.CODEC, PillarPlacementRules.simpleStoneOnlyStop(Blocks.REDSTONE_ORE));
+    public static final Feature<PillarFeatureConfiguration> LARGE_COAL_ORE_PILLAR =
+            new ConfigurableDripstoneStylePillarFeature(PillarFeatureConfiguration.CODEC, PillarPlacementRules.simpleStoneOnlyStop(Blocks.COAL_ORE));
 }